handle common reports

+++ corehq/apps/reports/management/commands/clean_orphaned_notifications.py
@@ -3,6 +3,7 @@ from __future__ import unicode_literals
 from django.core.management import BaseCommand
 from corehq.apps.reports.models import ReportNotification, ReportConfig
 from corehq.apps.userreports.models import ReportConfiguration
+from corehq.util.couch import get_document_or_not_found, DocumentNotFound
 from collections import defaultdict
 from couchdbkit.exceptions import ResourceNotFound
 from dimagi.utils.couch.undo import is_deleted
@@ -36,17 +37,22 @@ class Command(BaseCommand):
                 if not existent_config_ids[cid]:
                     try:
                         rc = ReportConfig.get(cid)
-                        rcuration = ReportConfiguration.get(rc.subreport_slug)
+                        if not rc.subreport_slug:  # seems to be the case of common reports
+                            existent_config_ids[cid] = True
+                            continue
+
+                        rcuration = get_document_or_not_found(
+                            ReportConfiguration, rc.domain, rc.subreport_slug)
                         if is_deleted(rcuration):
                             if options['execute']:
                                 super(type(rc), rc).delete()  # i am updating notifications manually
                             notification.config_ids.remove(cid)
                             updated_notification = True
-                        else:
-                            existent_config_ids[cid] = True
-                    except ResourceNotFound:
+                    except ResourceNotFound:  # ReportConfig not found
                         notification.config_ids.remove(cid)
                         updated_notification = True
+                    except DocumentNotFound:  # ReportConfiguration not found
+                        pass
 
             if updated_notification:
                 if options['execute']:
